Relocated history.tsx
